rag/requirements.txt
io
flask
PyPDF2
langchain.text_splitter
langchain_community.vectorstores